requirements.txt
asgiref==3.7.2
boto3==1.28.30
botocore==1.31.30
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
googlemaps==4.10.0
idna==3.4
jmespath==1.0.1
psycopg2-binary==2.9.7
pyshorteners==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
